+++ ntex/CHANGES.md
@@ -1,8 +1,16 @@
 # Changes
 
+## [3.0.0-pre.1] 2025-11-27
+
+* Update service definitions for ntex::web
+
 ## [3.0.0-pre.0] 2025-11-27
 
 * New io configuration subsystem
+
+* Move http client to top level
+
+* Update MSRV 1.85
 
 ## [2.17.0] 2025-11-12